# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dientes Fijos Medellín CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: view
          in: query
          schema:
            type: string
            enum: [day, week, month]
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "409":
          description: Time slot conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    put:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /appointments/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [appointments]
      summary: Get available time slots for a given date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: duration
          in: query
          schema:
            type: integer
            default: 60
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"
...
  /automations:
    get:
      operationId: listAutomations
      tags: [automations]
      summary: List automation rules
      responses:
        "200":
          description: List of automations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    post:
      operationId: createAutomation
      tags: [automations]
      summary: Create automation rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationBody"
      responses:
        "201":
          description: Automation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
  /automations/{id}:
    put:
      operationId: updateAutomation
      tags: [automations]
      summary: Update automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutomationBody"
      responses:
        "200":
          description: Automation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    delete:
      operationId: deleteAutomation
      tags: [automations]
      summary: Delete automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Automation deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        patientPhone:
          type: string
        treatment:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled, no_show]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, patientId, patientName, treatment, date, startTime, endTime, status, createdAt]
    CreateAppointmentBody:
      type: object
      properties:
        patientId:
          type: integer
        treatment:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        duration:
          type: integer
          default: 60
        notes:
          type: string
      required: [patientId, treatment, date, startTime]
    UpdateAppointmentBody:
      type: object
      properties:
        treatment:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        duration:
          type: integer
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled, no_show]
        notes:
          type: string
    TimeSlot:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        available:
          type: boolean
      required: [startTime, endTime, available]
...
    Automation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        trigger:
          type: string
          enum: [appointment_reminder, follow_up, new_patient_welcome, appointment_confirmed, missed_appointment, reactivation]
        message:
          type: string
        delayHours:
          type: integer
          nullable: true
        active:
          type: boolean
        executionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, trigger, message, active, executionCount, createdAt]
    CreateAutomationBody:
      type: object
      properties:
        name:
          type: string
        trigger:
          type: string
          enum: [appointment_reminder, follow_up, new_patient_welcome, appointment_confirmed, missed_appointment, reactivation]
        message:
          type: string
        delayHours:
          type: integer
        active:
          type: boolean
      required: [name, trigger, message]
    UpdateAutomationBody:
      type: object
      properties:
        name:
          type: string
        trigger:
          type: string
        message:
          type: string
        delayHours:
          type: integer
        active:
          type: boolean
...